query-scripts/requirements.txt
psycopg2
openpyxl
boto3
python-dotenv
pip